Working on the logging mechanism
