--- src/main/java/at/ac/fhcampuswien/richAF/data/LogManager.java
@@ -1,17 +1,0 @@
-package at.ac.fhcampuswien.richAF.data;
-
-public class LogManager {
-    private String fileName;
-
-    public LogManager(String fileName) {
-        this.fileName = fileName;
-    }
-
-    public void setFileName(String fileName) {
-        this.fileName = fileName;
-    }
-
-    public String getFileName() {
-        return fileName;
-    }
-}
